# openapi/connector_spec.yaml
openapi: 3.0.1
info:
  title: Connector API
  version: 1.0.0
paths:
...
  /v1/api/accounts/{accountId}/charges/{chargeId}/cancel:
    post:
      operationId: cancelCharge
      parameters:
      - description: Gateway account ID
        example: 1
        in: path
        name: accountId
        required: true
        schema:
          type: integer
          format: int64
      - description: Charge external ID
        example: b02b63b370fd35418ad66b0101
        in: path
        name: chargeId
        required: true
        schema:
          type: string
      responses:
        "202":
          description: Accepted - operation already in progress
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - charge is not in correct state
        "404":
          description: Not found - charge not found
      summary: Cancel charge
      tags:
      - Charge operations
  /v1/api/accounts/{accountId}/charges/{chargeId}/capture:
    post:
      description: "This endpoint should be called to capture a delayed capture charge.\
        \ The charge needs to have been previously marked as AWAITING CAPTURE REQUEST\
        \ for this call to succeed. When a charge is in any of the states CAPTURED,\
        \ CAPTURE APPROVED, CAPTURE APPROVED RETRY, CAPTURE READY, CAPTURE SUBMITTED\
        \ then nothing happens and the response will be a 204. When a charge is in\
        \ a status that cannot transition (eg. none of the above) then 409 response\
        \ is returned. "
      operationId: markChargeAsCaptureApproved
      parameters:
      - in: path
        name: accountId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: chargeId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No content
        "404":
          description: Not found - charge not found
        "409":
          description: Conflict - if charge is not in correct state
        "500":
          description: Internal server error
      summary: Mark delayed capture charge as eligible for capture and adds charge
        to capture queue
      tags:
      - Charge operations
...
  /v1/api/charges/authorise:
    post:
      operationId: authorise
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/AuthoriseRequest'
        required: true
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid one time token or one_time_token has
            already been used
        "402":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authorisation declined
        "404":
          description: Not found - charge not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity - Invalid payload or missing mandatory
            attributes
        "500":
          description: Authorisation error
      summary: Authorise MOTO (api) payment
      tags:
      - Charge operations
...
  /v1/frontend/charges/{chargeId}/3ds:
    post:
      operationId: authorise3dsCharge
      parameters:
      - description: Charge external ID
        example: b02b63b370fd35418ad66b0101
        in: path
        name: chargeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth3dsResult'
      responses:
        "200":
          description: OK
        "202":
          description: Accepted - payment has been submitted for 3ds authorisation
            and awaiting response from payment service provider
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid payload or the payment has been declined
        "402":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway error
        "404":
          description: Not found - charge not found
        "500":
          description: Internal server error - For gateway errors or anything else
            not handled
      summary: Authorise 3DS charge
      tags:
      - Charge operations
  /v1/frontend/charges/{chargeId}/cancel:
    post:
      operationId: userCancelCharge
      parameters:
      - description: Charge external ID
        example: b02b63b370fd35418ad66b0101
        in: path
        name: chargeId
        required: true
        schema:
          type: string
      responses:
        "202":
          description: Accepted - operation already in progress
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - charge is not in correct state
        "404":
          description: Not found - charge not found
      summary: Cancel charge (action by user)
      tags:
      - Charge operations
  /v1/frontend/charges/{chargeId}/capture:
    post:
      description: Marks charge as eligible (or ready - for delayed capture) for capture
        and also adds charge to capture queue (if not delayed capture).
      operationId: captureCharge
      parameters:
      - description: Charge external ID
        example: b02b63b370fd35418ad66b0101
        in: path
        name: chargeId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No content
        "400":
          description: Bad request - if charge is not in correct state
        "404":
          description: Not found - charge not found
        "500":
          description: Internal server error
      summary: Mark charge as eligible for capture
      tags:
      - Charge operations
  /v1/frontend/charges/{chargeId}/cards:
    post:
      operationId: authoriseCharge
      parameters:
      - description: Charge external ID
        example: b02b63b370fd35418ad66b0101
        in: path
        name: chargeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCardDetails'
      responses:
        "200":
          description: OK
        "202":
          description: Accepted - payment has been submitted for authorisation and
            awaiting response from payment service provider
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid payload or the payment has been declined
        "402":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway error
        "404":
          description: Not found - charge not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity - Invalid payload or missing mandatory
            attributes
        "500":
          description: Internal server error - For gateway errors or anything else
            not handled
      summary: Authorise charge
      tags:
      - Charge operations
...
  /v1/frontend/charges/{chargeId}/wallets/apple:
    post:
      operationId: authoriseApplePay
      parameters:
      - description: Charge external ID
        example: b02b63b370fd35418ad66b0101
        in: path
        name: chargeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplePayAuthRequest'
        required: true
      responses:
        "200":
          description: OK
        "202":
          description: Accepted - payment has been submitted for authorisation and
            awaiting response from payment service provider
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid payload or the payment has been declined
        "402":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway error
        "404":
          description: Not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity - Invalid payload or missing mandatory
            attributes
        "500":
          description: Internal server error
      summary: Authorise Apple Pay payment
      tags:
      - Charge operations
...
components:
  schemas:
...
    AuthoriseRequest:
      type: object
      properties:
        card_number:
          type: string
          example: "4242424242424242"
          maxLength: 19
          minLength: 12
        cardholder_name:
          type: string
          description: Cardholder name
          example: Joe B
          maxLength: 255
        cvc:
          type: string
          example: "123"
          maxLength: 4
          minLength: 3
        expiry_date:
          type: string
          description: 5 character string in MM/YY format
          example: 01/99
          maxLength: 5
          minLength: 5
        one_time_token:
          type: string
          description: the one time token provided in the `auth_url_post` link of
            the create payment API response
          example: 123abc123
      required:
      - card_number
      - cardholder_name
      - cvc
      - expiry_date
      - one_time_token
...
    ErrorResponse:
      type: object
      properties:
        error_identifier:
          type: string
          enum:
          - GENERIC
          - REFUND_NOT_AVAILABLE
          - REFUND_AMOUNT_AVAILABLE_MISMATCH
          - REFUND_NOT_AVAILABLE_DUE_TO_DISPUTE
          - ZERO_AMOUNT_NOT_ALLOWED
          - MOTO_NOT_ALLOWED
          - CANCEL_CHARGE_FAILURE_DUE_TO_CONFLICTING_TERMINAL_STATE_AT_GATEWAY_CHARGE_STATE_FORCIBLY_TRANSITIONED
          - CANCEL_CHARGE_FAILURE_DUE_TO_CONFLICTING_TERMINAL_STATE_AT_GATEWAY_INVALID_STATE_TRANSITION
          - AUTH_TOKEN_INVALID
          - AUTH_TOKEN_REVOKED
          - ACCOUNT_NOT_LINKED_WITH_PSP
          - TELEPHONE_PAYMENT_NOTIFICATIONS_NOT_ALLOWED
          - AGREEMENT_NOT_FOUND
          - ONE_TIME_TOKEN_INVALID
          - ONE_TIME_TOKEN_ALREADY_USED
          - INVALID_ATTRIBUTE_VALUE
          - CARD_NUMBER_REJECTED
          - AUTHORISATION_API_NOT_ALLOWED
          - AUTHORISATION_REJECTED
          - AUTHORISATION_ERROR
          - AUTHORISATION_TIMEOUT
          - AGREEMENT_NOT_ACTIVE
          - INCORRECT_AUTHORISATION_MODE_FOR_SAVE_PAYMENT_INSTRUMENT_TO_AGREEMENT
          - MISSING_MANDATORY_ATTRIBUTE
          - UNEXPECTED_ATTRIBUTE
          - ACCOUNT_DISABLED
          - RECURRING_CARD_PAYMENTS_NOT_ALLOWED
          - IDEMPOTENCY_KEY_USED
          - CARD_NUMBER_IN_PAYMENT_LINK_REFERENCE_REJECTED
          example: GENERIC
        message:
          type: array
          items:
            type: string
            example: error message
        reason:
          type: string
          example: "Optional - ex: amount_not_available"
...